@@ -3,8 +3,9 @@ package back.petionary.domain.account.entity;
 import back.petionary.common.BaseEntity;
 import back.petionary.common.util.validation.Validation;
 import java.time.LocalDate;
-import javax.persistence.Column;
-import javax.persistence.Entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
