@@ -1,25 +1,34 @@
 package back.petionary.domain.account.entity;
 
-import back.petionary.common.util.validation.Validation;
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
+import back.petionary.application.address.dto.AddressDetails;
+import back.petionary.common.BaseEntity;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
+
+import java.util.List;
 
 @NoArgsConstructor
 @Getter
 @Embeddable
-public class Address {
+@Entity
+public class Address extends BaseEntity {
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    private Long id;
 
-    @Column(nullable = false)
-    private String zipCode; //우편번호
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Account account;
 
-    @Column(nullable = false)
-    private String detail; //상세주소
+    private String area;
+    private String city;
+    private String localAddress;
 
-    public Address(String zipCode, String detail) {
-        Validation.validateNull(zipCode,detail);
-        this.zipCode = zipCode;
-        this.detail = detail;
+    public Address(Account account, String area, String city, String localAddress) {
+        this.account = account;
+        this.area = area;
+        this.city = city;
+        this.localAddress = localAddress;
     }
 }
